MyWebIntelligenceAPI/requirements.txt
# Framework principal
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Base de données
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.0

# ORM et requêtes async
psycopg2-binary==2.9.9

# Tâches asynchrones
celery==5.3.4
redis==5.0.1
flower==2.0.1  # Pour le monitoring Celery

# HTTP & Scraping
httpx==0.25.2
beautifulsoup4==4.12.2
lxml==4.9.3
aiohttp==3.9.1

# Authentification & Sécurité
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Traitement de contenu
trafilatura==1.6.3  # Pour l'extraction de contenu
newspaper3k==0.2.8  # Alternative pour l'extraction
readability-lxml==0.8.1  # Extraction de contenu lisible

# Analyse de médias
pillow==10.1.0  # Traitement d'images
python-magic==0.4.27  # Détection de types de fichiers
imagehash==4.3.1  # Hash perceptuel des images
colorthief==0.2.1  # Extraction de couleurs dominantes
exifread==3.0.0  # Lecture des métadonnées EXIF

# WebSocket
websockets==12.0

# Monitoring & Logging
prometheus-client==0.12.0
starlette-prometheus==0.9.0
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2  # Pour les tests d'API
pytest-cov==4.1.0
factory-boy==3.3.0  # Pour les fixtures de test

# Export de données
pandas==2.1.4
networkx==3.2.1  # Pour GEXF
python-igraph==0.11.3  # Graphes alternatifs

# Utilitaires
email-validator==2.1.1
python-dotenv==1.0.0
pytz==2023.3
pyyaml==6.0.1

# Archive.org support
internetarchive==3.6.0

# Playwright pour scraping dynamique (optionnel)
# playwright==1.40.0  # À installer séparément avec: playwright install
